Need to update the document: Cannot add hostname for -RemoteAddress in New-NetFirewall
The -RemoteAddress parameter specifies this parameter value can be an IPv4 or IPv6 address, hostname, subnet, range, or the following keyword: Any. However the acceptable formats for this parameter are only address, subnet or a range. We cannot add hostname for this parameter.
I request you to change this public documentation.

One of our Microsoft customers requires this change to be made and to provide him a documentation for this change.